crypto_screener_ai/requirements.txt
requests
python-binance
numpy
pandas
ta
